refactor: split health controller into single-action controllers
Replace the single HealthController with four focused controllers
(HealthController, LiveController, MetricsController, GroupsController),
each exposing only a #show action. Move shared run_checks to
ApplicationController. Routes and endpoints are unchanged.
