+++ app/controllers/rails_health_checks/application_controller.rb
@@ -4,5 +4,13 @@ module RailsHealthChecks
   class ApplicationController < ActionController::API
     include Authentication
     before_action :authenticate!
+
+    private
+
+    def run_checks(check_names)
+      config = RailsHealthChecks.configuration
+      checks = CheckRegistry.build(check_names)
+      CheckRegistry.run(checks, timeout: config.timeout)
+    end
   end
 end
